+++ src/remote/activitypub/models/person.ts
@@ -14,7 +14,7 @@ import { extractApHashtags } from './tag';
 import { apLogger } from '../logger';
 import { Note } from '../../../models/entities/note';
 import { updateUsertags } from '../../../services/update-hashtag';
-import { Users, Instances, DriveFiles, Followings, UserProfiles, UserPublickeys } from '../../../models';
+import { Users, Instances, Followings, UserProfiles, UserPublickeys } from '../../../models';
 import { User, IRemoteUser } from '../../../models/entities/user';
 import { Emoji } from '../../../models/entities/emoji';
 import { UserNotePining } from '../../../models/entities/user-note-pinings';
@@ -33,6 +33,7 @@ import { resolveUser } from '../../resolve-user';
 import { StatusError } from '@/misc/fetch';
 import { MAX_NAME_LENGTH, MAX_SUMMARY_LENGTH } from '../../../misc/hard-limits';
 import { truncate } from '@/misc/truncate';
+import { resolveAnotherUser } from '../resolve-another-user';
 
 const logger = apLogger;
 
@@ -141,6 +142,14 @@ export async function createPerson(uri: string, resolver?: Resolver): Promise<Us
 
 	const bday = person['vcard:bday']?.match(/^[0-9]{4,8}-\d{2}-\d{2}/);
 
+	const movedTo = (person.id && person.movedTo)
+		? await resolveAnotherUser(person.id, person.movedTo, resolver)
+			.catch(e => {
+				logger.warn(`Error in movedTo: ${e}`);
+				return null;
+			})
+		: null;
+
 	// Create user
 	let user: IRemoteUser;
 	try {
@@ -166,7 +175,8 @@ export async function createPerson(uri: string, resolver?: Resolver): Promise<Us
 				tags,
 				isBot,
 				isCat: (person as any).isCat === true,
-				isLady: (person as any).isLady === true
+				isLady: (person as any).isLady === true,
+				movedToUserId: movedTo?.id || null,
 			}).then(x => transactionalEntityManager.findOneOrFail(User, x.identifiers[0])) as IRemoteUser;
 
 			await transactionalEntityManager.insert(UserProfile, {
@@ -236,26 +246,14 @@ export async function createPerson(uri: string, resolver?: Resolver): Promise<Us
 
 	const avatarId = avatar ? avatar.id : null;
 	const bannerId = banner ? banner.id : null;
-	const avatarUrl = avatar ? DriveFiles.getPublicUrl(avatar, true) : null;
-	const bannerUrl = banner ? DriveFiles.getPublicUrl(banner) : null;
-	const avatarBlurhash = avatar ? avatar.blurhash : null;
-	const bannerBlurhash = banner ? banner.blurhash : null;
 
 	await Users.update(user!.id, {
 		avatarId,
 		bannerId,
-		avatarUrl,
-		bannerUrl,
-		avatarBlurhash,
-		bannerBlurhash
 	});
 
 	user!.avatarId = avatarId;
 	user!.bannerId = bannerId;
-	user!.avatarUrl = avatarUrl;
-	user!.bannerUrl = bannerUrl;
-	user!.avatarBlurhash = avatarBlurhash;
-	user!.bannerBlurhash = bannerBlurhash;
 	//#endregion
 
 	//#region カスタム絵文字取得
@@ -331,6 +329,14 @@ export async function updatePerson(uri: string, resolver?: Resolver | null, hint
 
 	const bday = person['vcard:bday']?.match(/^[0-9]{4,8}-\d{2}-\d{2}/);
 
+	const movedTo = (person.id && person.movedTo)
+		? await resolveAnotherUser(person.id, person.movedTo, resolver)
+			.catch(e => {
+				logger.warn(`Error in movedTo: ${e}`);
+				return null;
+			})
+		: null;
+
 	const updates = {
 		lastFetchedAt: new Date(),
 		inbox: person.inbox,
@@ -345,18 +351,15 @@ export async function updatePerson(uri: string, resolver?: Resolver | null, hint
 		isLady: (person as any).isLady === true,
 		isLocked: !!person.manuallyApprovesFollowers,
 		isExplorable: !!person.discoverable,
+		movedToUserId: movedTo?.id || null,
 	} as Partial<User>;
 
 	if (avatar) {
 		updates.avatarId = avatar.id;
-		updates.avatarUrl = DriveFiles.getPublicUrl(avatar, true);
-		updates.avatarBlurhash = avatar.blurhash;
 	}
 
 	if (banner) {
 		updates.bannerId = banner.id;
-		updates.bannerUrl = DriveFiles.getPublicUrl(banner);
-		updates.bannerBlurhash = banner.blurhash;
 	}
 
 	// Update user